% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDD.R
\name{center}
\alias{center}
\title{Center a Running Variable}
\usage{
center(df, var, center = 0)
}
\arguments{
\item{df}{a data frame}

\item{var}{the variable you wish to center}

\item{center}{the value around which you wish to center your variable (default is zero--does nothing)}
}
\value{
a data frame with the chosen variable centered
}
\description{
Center a Running Variable
}
