% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{coinflips}
\alias{coinflips}
\title{Flip a coin a few times}
\usage{
coinflips(n = 10000, m = 100)
}
\arguments{
\item{n}{how many trials to run}

\item{m}{how many coinflips per trial}
}
\description{
from drob's twitter
https://twitter.com/drob/status/1008409373423611904
}
