% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{correlate}
\alias{correlate}
\title{Correlation}
\usage{
correlate(df, x, y, ...)
}
\arguments{
\item{df}{A data frame}

\item{x}{The x variable}

\item{y}{The y variable}
}
\description{
A fixed correlation function; it should operate in a pipe and also defaults to
pairwise complete observations being true. Additional arguments to base `cor()`
are also available.
}
