% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{description}
\alias{description}
\title{Description of All variables in a Dataset}
\usage{
description(data, group = NULL, fast = TRUE, ...)
}
\arguments{
\item{data}{your data}

\item{group}{optional: your grouping variable}
}
\value{
an HTML object in the viewer pane (or in a shiny app or markdown document)
}
\description{
cleans the output of `psych::describe()` and passes it into
a `gt::gt()` object to make an HTML output
}
