% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{diagnostics}
\alias{diagnostics}
\title{Model Diagnostics}
\usage{
diagnostics(model)
}
\arguments{
\item{model}{the linear model}
}
\description{
Of unknown provenance--been in my files for years now.
Returns a model sumamry plus three graphs of model fit statistics/diagnostics
in base R.
}
