% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{dossier}
\alias{dossier}
\title{Print a Dossier}
\usage{
dossier(df, id, value, ...)
}
\arguments{
\item{df}{a dataframe}

\item{id}{the id variable of your data frame}

\item{value}{the value of your individual dossier (must be in quotations)}
}
\description{
Takes an input of a specific id for a specific observation and returns the name and value for each variable in the dataset for that single observation. A different way of examining what your dataset looks like on a micro level.
}
