% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata.R
\name{drop}
\alias{drop}
\title{Drop an object}
\usage{
drop(df)
}
\arguments{
\item{df}{a data frame (or any other object)}
}
\description{
For when you forget that you're workign in R and use a STATA command instead.
}
