% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_plots.R
\name{gg_added_var}
\alias{gg_added_var}
\title{Create an Added Variable Plot}
\usage{
gg_added_var(partial, extended, se = TRUE)
}
\arguments{
\item{partial}{the model containing only the output and the variable for which you wish to graph the av plot}

\item{extended}{the full model from which you are drawing the partial variable}

\item{se}{whether to graph the standard error around the regression line}
}
\value{
a ggplot2 graph that can be extended via other ggplot2 arguments
}
\description{
Adapted from Steven Pollack
https://github.com/stevenpollack/stat151a/blob/master/From_Lab/Feb-26-2014.R
(combines his data manipulation and graphing of av plots and removes formatting (can be added back in or adjusted as needed)
}
