% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{is_extant}
\alias{is_extant}
\title{Check if a variable has data extant}
\usage{
is_extant(x)
}
\arguments{
\item{x}{a data vector}
}
\value{
a logical: TRUE or FALSE
}
\description{
Sometimes for longitudinal data, variables exist but have not been entered yet. To ensure forward compatibility, this function checks to see if any data have yet been entered for a given variable.
}
\details{
works best with "select_if()" as in:
data %>% select_if(is_extant)
}
