% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{is_numeric}
\alias{is_numeric}
\title{Check if a variable is Numeric}
\usage{
is_numeric(x)
}
\arguments{
\item{x}{a data vector}
}
\value{
a logical: TRUE or FALSE
}
\description{
Sometimes you have an unknown number of variables being piped into a function that only takes numeric vectors. This function exists to select only those which will work in the final function.
}
\details{
works best with "select_if()" as in:
data %>% select_if(is_numeric)
}
