% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{palprint}
\alias{palprint}
\title{Print a palette of colors}
\usage{
palprint(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{a color palette}

\item{n}{how many colors to select}

\item{type}{whether to print the colors as discrete or continuous}
}
\description{
Maybe stolen from Karthik Ram?
}
