% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{paste_data}
\alias{paste_data}
\title{Paste data from an Excel Spreadsheet
from https://www.r-bloggers.com/copying-data-from-excel-to-r-and-back/
adapted for tibbles}
\usage{
paste_data(header = TRUE, ...)
}
\value{
a tibble of the pasted data
}
\description{
Paste data from an Excel Spreadsheet
from https://www.r-bloggers.com/copying-data-from-excel-to-r-and-back/
adapted for tibbles
}
