% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{plot_freq}
\alias{plot_freq}
\title{Plot Factors by Frequency}
\usage{
plot_freq(data, group, n = 10)
}
\arguments{
\item{data}{a data frame}

\item{group}{a variable of factors}

\item{n}{how many to print into the ggplot graph}
}
\value{
a ggplot graph object
}
\description{
https://stackoverflow.com/questions/46862482/plot-a-descending-frequency-bar-chart-using-a-custom-function-with-ggplot2-dply
}
