% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata.R
\name{regress}
\alias{regress}
\title{Summarize a Regression}
\usage{
regress(df, ...)
}
\arguments{
\item{df}{a data frame}
}
\value{
A regression summary
}
\description{
Reorders summarize(lm()) to allow it to be the last verb of a pipe and operate in one step.
}
