% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_color_inova}
\alias{scale_color_inova}
\title{Scale Colors as inova healthsystems}
\usage{
scale_color_inova(palette = c("inova"), alpha = 1, ...)
}
\description{
Scale Colors as inova healthsystems
}
