% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{solve_quadratic}
\alias{solve_quadratic}
\title{Solve a Quadratic Equation}
\usage{
solve_quadratic(a, b, c)
}
\arguments{
\item{a}{the first coefficient in standard form}

\item{b}{the second coefficient}

\item{c}{the third coefficient}
}
\value{
two roots either real or imaginary. They will be in complex form notation. Roots are real if the imaginary term is 0i. If one term is 0 + 0i then there is only one root.
}
\description{
Just another practice function.
returns roots in complex notation regardless of root form
}
