% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata.R
\name{stata_browse}
\alias{stata_browse}
\title{Browse a Data Frame}
\usage{
stata_browse(df)
}
\arguments{
\item{df}{a data frame}
}
\description{
For when you forget that you're workign in R and use a STATA command instead.
}
