% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata.R
\name{stata_summary}
\alias{stata_summary}
\title{Output Regression Summary in Stata Format}
\usage{
stata_summary(x, ...)
}
\arguments{
\item{x}{A regression Model}
}
\description{
twitter post deleted, author lost to history
}
