% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata.R
\name{summarize}
\alias{summarize}
\title{Summarize a Data Frame}
\usage{
summarize(df, ...)
}
\arguments{
\item{df}{a data frame}
}
\description{
For when you forget that you're workign in R and use a STATA command instead.
}
