% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata.R
\name{use}
\alias{use}
\title{Import a data set}
\usage{
use(name)
}
\arguments{
\item{name}{the file-name of the data you wish to import}
}
\description{
For when you forget that you're workign in R and use a STATA command instead.
}
