% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{vlookup}
\alias{vlookup}
\title{Vlookup a value}
\usage{
vlookup(this, data, key, value)
}
\arguments{
\item{this}{a list of values}

\item{data}{a dataset to match it to}

\item{key}{the variable in the data to match on}

\item{value}{the value applied to "this"}
}
\description{
from Jenny Bryan @JennyBryan
https://twitter.com/JennyBryan/status/980978609794895872
}
