% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{wrapper}
\alias{wrapper}
\title{Wrap text in ggplot2}
\usage{
wrapper(x, ...)
}
\arguments{
\item{x}{some text for a ggplot label.}
}
\value{
wrapped text for a ggplot label.
}
\description{
taken from twitter @geokaramanis but he attributes it to an anon stack overflow postl
}
\details{
example: subtitle = wrapper("Really long text...", width = 80)
}
