% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccrr.R
\name{plot_model}
\alias{plot_model}
\title{Plot a simple linear model}
\usage{
plot_model(mod, explanatory, response, .fitted = ".fitted")
}
\arguments{
\item{mod}{the linear model}

\item{explanatory}{the x variable}

\item{response}{the y variable}
}
\description{
https://twitter.com/katiejolly6/status/960653271080865798
http://katiejolly.io/blog/2018-02-05/aes-string
}
